More accurate owner information and cross-library compatibility tests
This change makes reporting more accurate when a member is found on a supertype.

For example, if `A` tries to call `B.foo()` which is defined on `B`'s superclass `C` as `private foo()`, the tool will now report inaccessible member `C.foo` instead of `B.foo`. 

Also adding a test that covers the above and exercises cross-library compatibility checking.