Add the concept of roots
In larger projects, a lot of noise is produced by unused classes in
dependencies. To deal with some of that noise, we add the concept of
"roots," which is a subset of application classes. Instead of validating
all application classes, we start with roots and recursively traverse
application classes reachable from those roots. The resulting closure is
the set of application classes that we will actually validate. By
default, the roots are all classes in the current project.